+++ src/lottoConstants/systemConstants.js
@@ -1,6 +1,6 @@
 export const LOTTO_NUMBERS = {
   LENGTH: 6,
-  BONUS_LEGNTH: 1,
+  BONUS_LENGTH: 1,
   MIN: 1,
   MAX: 45,
 };
