+++ style.css
@@ -168,6 +168,10 @@ button {
   width: 1rem;
   height: 2.2rem;
 }
+.bonus-input {
+  width: 1rem;
+  height: 2.2rem;
+}
 .bonus-input-container {
   display: flex;
   flex-direction: column;
